requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
Faker==37.1.0
python-dotenv==0.21.0
joblib==1.4.2 
tensorflow==2.19.0
keras==3.9.2
